Pulling out sending functionality from FASubmissionFull to an abstract Sendable class
